add classic MDP
