+++ place_env/__init__.py
@@ -1,6 +1,11 @@
 from gym.envs.registration import register
 
 register(
-    id='Place-v0',
-    entry_point='place_env.placement:Placement',
+    id='Swap-v0',
+    entry_point='place_env.swapMDP:SwapPlacement',
+)
+
+register(
+    id='Classic-v0',
+    entry_point='place_env.classicMDP:ClassicPlacement'
 )
